requirements.txt
google-adk==1.28.0
google-genai==1.70.0
python-docx==1.2.0
python-dotenv==1.2.2
pdfplumber==0.11.9
fastapi==0.135.3
uvicorn==0.34.0
asyncpg==0.30.0
